KEYCLOAK-18635 Move classes specific to ConcurrentHashMap to chm package
