+++ model/map/src/main/java/org/keycloak/models/map/storage/chm/MapModelCriteriaBuilder.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.keycloak.models.map.storage;
+package org.keycloak.models.map.storage.chm;
 
 import org.keycloak.models.map.common.AbstractEntity;
 import org.keycloak.storage.SearchableModelField;
@@ -22,6 +22,7 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
+import org.keycloak.models.map.storage.ModelCriteriaBuilder;
 
 /**
  *
